requirements.txt
discord.py-self
python-dotenv
pytz
asyncpg
